Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/imagebitmap/createImageBitmap-invalid-args-expected.txt


PASS createImageBitmap with an HTMLCanvasElement source and sw set to 0
PASS createImageBitmap with an HTMLCanvasElement source and sh set to 0
FAIL createImageBitmap with an HTMLCanvasElement source and oversized (unallocatable) crop region assert_equals: expected 100000000 but got 20
PASS createImageBitmap with an HTMLVideoElement source and sw set to 0
PASS createImageBitmap with an HTMLVideoElement source and sh set to 0
FAIL createImageBitmap with an HTMLVideoElement source and oversized (unallocatable) crop region assert_equals: expected 100000000 but got 20
PASS createImageBitmap with an HTMLVideoElement from a data URL source and sw set to 0
PASS createImageBitmap with an HTMLVideoElement from a data URL source and sh set to 0
FAIL createImageBitmap with an HTMLVideoElement from a data URL source and oversized (unallocatable) crop region assert_equals: expected 100000000 but got 20
PASS createImageBitmap with a bitmap HTMLImageElement source and sw set to 0
PASS createImageBitmap with a bitmap HTMLImageElement source and sh set to 0
FAIL createImageBitmap with a bitmap HTMLImageElement source and oversized (unallocatable) crop region assert_equals: expected 100000000 but got 20
PASS createImageBitmap with a vector HTMLImageElement source and sw set to 0
PASS createImageBitmap with a vector HTMLImageElement source and sh set to 0
FAIL createImageBitmap with a vector HTMLImageElement source and oversized (unallocatable) crop region assert_equals: expected 100000000 but got 20
FAIL createImageBitmap with a bitmap SVGImageElement source and sw set to 0 promise_rejects_js: function "function () { throw e }" threw object "TypeError: Type error" ("TypeError") expected instance of function "function RangeError() {
    [native code]
}" ("RangeError")
FAIL createImageBitmap with a bitmap SVGImageElement source and sh set to 0 promise_rejects_js: function "function () { throw e }" threw object "TypeError: Type error" ("TypeError") expected instance of function "function RangeError() {
    [native code]
}" ("RangeError")
FAIL createImageBitmap with a bitmap SVGImageElement source and oversized (unallocatable) crop region assert_throws_dom: function "() => { throw e }" threw object "TypeError: Type error" that is not a DOMException InvalidStateError: property "code" is equal to undefined, expected 11
FAIL createImageBitmap with a vector SVGImageElement source and sw set to 0 promise_rejects_js: function "function () { throw e }" threw object "TypeError: Type error" ("TypeError") expected instance of function "function RangeError() {
    [native code]
}" ("RangeError")
FAIL createImageBitmap with a vector SVGImageElement source and sh set to 0 promise_rejects_js: function "function () { throw e }" threw object "TypeError: Type error" ("TypeError") expected instance of function "function RangeError() {
    [native code]
}" ("RangeError")
FAIL createImageBitmap with a vector SVGImageElement source and oversized (unallocatable) crop region assert_throws_dom: function "() => { throw e }" threw object "TypeError: Type error" that is not a DOMException InvalidStateError: property "code" is equal to undefined, expected 11
FAIL createImageBitmap with an OffscreenCanvas source and sw set to 0 promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: OffscreenCanvas"
FAIL createImageBitmap with an OffscreenCanvas source and sh set to 0 promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: OffscreenCanvas"
FAIL createImageBitmap with an OffscreenCanvas source and oversized (unallocatable) crop region promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: OffscreenCanvas"
PASS createImageBitmap with an ImageData source and sw set to 0
PASS createImageBitmap with an ImageData source and sh set to 0
FAIL createImageBitmap with an ImageData source and oversized (unallocatable) crop region assert_equals: expected 100000000 but got 20
PASS createImageBitmap with an ImageBitmap source and sw set to 0
PASS createImageBitmap with an ImageBitmap source and sh set to 0
FAIL createImageBitmap with an ImageBitmap source and oversized (unallocatable) crop region assert_equals: expected 100000000 but got 20
PASS createImageBitmap with a Blob source and sw set to 0
PASS createImageBitmap with a Blob source and sh set to 0
FAIL createImageBitmap with a Blob source and oversized (unallocatable) crop region assert_equals: expected 100000000 but got 20
PASS createImageBitmap with undefined image source.
PASS createImageBitmap with null image source.
PASS createImageBitmap with CanvasRenderingContext2D image source.
PASS createImageBitmap with WebGLRenderingContext image source.
PASS createImageBitmap with Uint8Array image source.
PASS createImageBitmap with ArrayBuffer image source.
PASS createImageBitmap with empty image source.
PASS createImageBitmap with empty video source.
PASS createImageBitmap with an oversized canvas source.
FAIL createImageBitmap with an invalid OffscreenCanvas source. promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: OffscreenCanvas"
PASS createImageBitmap with an undecodable blob source.
PASS createImageBitmap with a broken image source.
FAIL createImageBitmap with an available but undecodable image source. assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL createImageBitmap with an available but zero height image source. assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL createImageBitmap with an available but zero width image source. assert_unreached: Should have rejected: undefined Reached unreachable code
PASS createImageBitmap with a closed ImageBitmap.
